[Refactor] #129 - RegexView 리펙토링 완료
