+++ PennyPack/Views/ScanView/ScanComponent/RegexHelper.swift
@@ -1,0 +1,42 @@
+import Foundation
+
+let testPrices = ["12.34", "€123.45", ".33€", ",88€", "7.89€/Kg", "4.33€/Kg"]
+let testItems = ["soupe à l'oignon", "1.30€/Kg", "234729", "500015740722", "0.84", "43/-/A/ ----"]
+
+let validT = extractValidPrices(testPrices)
+let validT2 = extractValidItems(testItems)
+
+
+func extractValidItems(_ strings: [String]) -> [String] {
+    let itemPattern = "[a-zA-ZàâäæáãåāèéêëęėēîïīįíìôōøõóòöœùûüūúÿçćčńñÀÂÄÆÁÃÅĀÈÉÊËĘĖĒÎÏĪĮÍÌÔŌØÕÓÒÖŒÙÛÜŪÚŸÇĆČŃÑ]+"
+    let regex = try! NSRegularExpression(pattern: itemPattern, options: .caseInsensitive)
+    
+    return strings.compactMap { string in
+        let filteredString = string.filter {
+            $0.isLetter
+        }
+        guard filteredString.count >= 4 else {
+            return nil
+        }
+        let range = NSRange(location:0, length: string.utf16.count)
+        guard regex .firstMatch(in: string, options: [], range: range) != nil else {
+            return nil
+        }
+        return string
+    }
+}
+
+func extractValidPrices(_ strings: [String]) -> [Double] {
+    let pricePattern = #"^€?\d*[.,]\d{1,2}€?$"#
+    let regex = try! NSRegularExpression(pattern: pricePattern)
+    
+    return strings.compactMap { string in
+        let range = NSRange(location: 0, length: string.utf16.count)
+        guard regex.firstMatch(in: string, options: [], range: range) != nil else {
+            return nil
+        }
+        let numberString = string.replacingOccurrences(of: "€", with: "")
+            .replacingOccurrences(of: ",", with: ".")
+        return Double(numberString)
+    }
+}
